+++ benchmarks/sindex.py
@@ -27,12 +27,11 @@ def generate_test_df():
     }
     # ensure index is pre-generated
     for data_type in data.keys():
-        data[data_type].sindex.query(data[data_type].geometry.values.data[0])
+        data[data_type].sindex.query(data[data_type].geometry.values[0])
     return data
 
 
 class BenchIntersection:
-
     param_names = ["input_geom_type", "tree_geom_type"]
     params = [
         geom_types,
@@ -54,7 +53,6 @@ class BenchIntersection:
 
 
 class BenchIndexCreation:
-
     param_names = ["tree_geom_type"]
     params = [
         geom_types,
@@ -77,11 +75,10 @@ class BenchIndexCreation:
         tree = self.data[tree_geom_type].sindex
         # also do a single query to ensure the index is actually
         # generated and used
-        tree.query(self.data[tree_geom_type].geometry.values.data[0])
+        tree.query(self.data[tree_geom_type].geometry.values[0])
 
 
 class BenchQuery:
-
     param_names = ["predicate", "input_geom_type", "tree_geom_type"]
     params = [
         predicates,
@@ -93,12 +90,12 @@ class BenchQuery:
         self.data = generate_test_df()
 
     def time_query_bulk(self, predicate, input_geom_type, tree_geom_type):
-        self.data[tree_geom_type].sindex.query_bulk(
-            self.data[input_geom_type].geometry.values.data,
+        self.data[tree_geom_type].sindex.query(
+            self.data[input_geom_type].geometry.values,
             predicate=predicate,
         )
 
     def time_query(self, predicate, input_geom_type, tree_geom_type):
         tree = self.data[tree_geom_type].sindex
-        for geom in self.data[input_geom_type].geometry.values.data:
+        for geom in self.data[input_geom_type].geometry.values:
             tree.query(geom, predicate=predicate)